animated player / commented out projectile
I commented out projectile since we're using punches. We can keep this in for further iterations. I think melee attacks will create a "projectile" in the direction of the punch, which will be static and disappear quickly. This'll be a hitbox. In otherwords, melee attacks will use a lot of the same code that projectiles do.